Jenkins container
Plugins are installed using first startup.
Default jobs are moved on first startup.
When running everything is persisted into dev-jenkins/data folder.
Compiling Golang tools to be used in Jenkins for SSH (and other complex) operations.

dev-jenkins/data need to be created with user permissions (not root).
docker-compose does not support bash interpolation, so default Ubuntu user is is used.